pyproject.toml
[project]
name = "data-platform"
version = "0.1.0"
description = "The data platform for American Beauty Institute."
authors = [
    {name = "Kim, Andrew", email = "andrewkimka@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "dagster>=1.12.22",
    "dagster-gcp>=0.28.22",
    "dagster-sqlmesh>=0.22.0",
    "facebook-business>=25.0.1",
    "fastapi>=0.115.0",
    "google-ads>=28.2.0",
    "google-analytics-data>=0.19.0",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-secret-manager>=2.27.0",
    "google-cloud-storage>=2.19.0",
    "pyarrow>=21.0.0",
    "pydantic-settings>=2.9.0",
    "python-dotenv>=1.1.1",
    "sqlmesh[bigquery]>=0.233.0",
    "stripe>=15.0.1",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "coverage>=7.9.1",
    "dagster-webserver>=1.12.22",
    "httpx>=0.28.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.0",
    "ty>=0.0.29",
]
docs = [
    "mkdocs-material>=9.6.22",
]

[tool.coverage.run]
branch = true
source = ["data-platform"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = []

[tool.ruff]
target-version = "py311"
line-length = 88
fix = true
unsafe-fixes = true

lint.select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
  "B",   # bugbear
  "SIM", # simplify
  "D",   # pydocstyle
  "PL",  # pylint-like
  "C90", # mccabe
  "PERF", # performance
  "RUF", # ruff-specific
  "PTH", # pathlib
  "FLY", # f-strings
]

lint.ignore = []

exclude = ["tests/data", ".venv", "build", "dist"]

[tool.ruff.lint.pydocstyle]
convention = "google"  # or "numpy" or "pep257"
